requirements.txt
python-telegram-bot[asyncio]
python-dotenv
aiohttp
pymysql
requests
mysql-connector-python
aiomysql